281: relocate contents of interpret/interpret.js
