+++ src/global/determinant.js
