+++ src/interpret/test/interpret.spec.js
@@ -1,6 +1,6 @@
 import test from 'tape';
 
-import { determinant } from '../interpret';
+import { determinant } from '../../global/determinant';
 
 test('determinant returns correct value for a 3x3', t => {
   t.is(determinant([
